+++ test/DefaultPoolTest.php
@@ -4,7 +4,7 @@ namespace Amp\Postgres\Test;
 
 use Amp\Delayed;
 use Amp\Loop;
-use Amp\Postgres\DefaultPool;
+use Amp\Postgres\Pool;
 use Amp\Postgres\ConnectionConfig;
 use PHPUnit\Framework\TestCase;
 
@@ -16,13 +16,13 @@ class DefaultPoolTest extends TestCase
      */
     public function testInvalidMaxConnections()
     {
-        new DefaultPool(new ConnectionConfig('connection string'), 0);
+        new Pool(new ConnectionConfig('connection string'), 0);
     }
 
     public function testIdleConnectionsRemovedAfterTimeout()
     {
         Loop::run(function () {
-            $pool = new DefaultPool(new ConnectionConfig('host=localhost user=postgres'));
+            $pool = new Pool(new ConnectionConfig('host=localhost user=postgres'));
             $pool->setIdleTimeout(2);
             $count = 3;
