+++ test/PgSqlPoolTest.php
@@ -4,7 +4,7 @@ namespace Amp\Postgres\Test;
 
 use Amp\Postgres\ConnectionConfig;
 use Amp\Postgres\Link;
-use Amp\Postgres\DefaultPool;
+use Amp\Postgres\Pool;
 use Amp\Promise;
 use Amp\Sql\Connector;
 use Amp\Success;
@@ -36,7 +36,7 @@ class PgSqlPoolTest extends AbstractLinkTest
                 return new Success();
             }));
 
-        $pool = new DefaultPool(new ConnectionConfig('connection string'), \count($this->handles), $connector);
+        $pool = new Pool(new ConnectionConfig('connection string'), \count($this->handles), $connector);
 
         $handle = \reset($this->handles);
