+++ test/PooledStatementTest.php
@@ -6,9 +6,9 @@ use Amp\Delayed;
 use Amp\Loop;
 use Amp\PHPUnit\TestCase;
 use Amp\Postgres\ConnectionConfig;
-use Amp\Postgres\Internal\PooledStatement;
-use Amp\Postgres\DefaultPool;
+use Amp\Postgres\Pool;
 use Amp\Postgres\ResultSet;
+use Amp\Sql\PooledStatement;
 use Amp\Sql\Statement;
 use Amp\Success;
 
@@ -17,7 +17,7 @@ class PooledStatementTest extends TestCase
     public function testActiveStatementsRemainAfterTimeout()
     {
         Loop::run(function () {
-            $pool = new DefaultPool(new ConnectionConfig('host=localhost user=postgres'));
+            $pool = new Pool(new ConnectionConfig('host=localhost user=postgres'));
 
             $statement = $this->createMock(Statement::class);
             $statement->method('getQuery')
@@ -44,7 +44,7 @@ class PooledStatementTest extends TestCase
     public function testIdleStatementsRemovedAfterTimeout()
     {
         Loop::run(function () {
-            $pool = new DefaultPool(new ConnectionConfig('host=localhost user=postgres'));
+            $pool = new Pool(new ConnectionConfig('host=localhost user=postgres'));
 
             $statement = $this->createMock(Statement::class);
             $statement->method('getQuery')
